big refactor. i wanted to seprate embedapi v1 and v2. embedapi v1 working with baloo iframe. i dint test it but i think is it good. need to ask about this i n function bind
