+++ src/lib/ekoPlayerPublicApi/v2.js
@@ -1,0 +1,37 @@
+function play() {
+
+}
+
+function pause() {
+
+}
+
+function load() {
+
+}
+
+function invoke() {
+
+}
+
+function on() {
+
+}
+
+function once() {
+
+}
+
+function off() {
+
+}
+
+export default {
+    load,
+    play,
+    pause,
+    invoke,
+    on,
+    once,
+    off
+};
